Add files via upload
Screen Setup
The turtle graphics window is created with a specified title and dimensions. This window is where the flag will be drawn.
Creating the Turtle Object
A turtle object is instantiated to act as the "pen" that will draw the flag on the screen. The speed of the turtle's movement is set to a moderate level.
Drawing Rectangles
A function is defined to draw a filled rectangle of a specified color, width, and height. This function will be used to draw the three horizontal stripes of the flag.
Positioning the Turtle
The turtle is moved to the starting position for drawing the first (saffron) stripe.
Drawing the Flag Stripes
The saffron, white, and green stripes of the flag are drawn one by one. After drawing each stripe, the turtle is repositioned to the starting point for the next stripe.
Drawing the Ashoka Chakra
A function is defined to draw the Ashoka Chakra, which is a navy blue wheel with 24 spokes located in the center of the white stripe.
The turtle is moved to the center position of the white stripe.
The outer circle of the Ashoka Chakra is drawn.
The 24 spokes are drawn by repeatedly moving the turtle forward from the center to the edge of the circle and then back to the center, rotating 15 degrees each time.
Finishing Up
The turtle is hidden to clean up the final appearance of the drawing.
The turtle graphics window is kept open so that the drawing remains visible until the user closes the window.
Use of the Code
This code provides a visual representation of the Indian national flag using Python's turtle graphics. It can be used as an educational tool to demonstrate programming concepts, graphical drawing techniques, or to create a graphical representation of the flag for various applications.